waxer: Wikimedia Analytics Query Service (AQS) API

This is an R wrapper for the Wikimedia Analytics Query Service (AQS). The REST API provides cacheable and straightforward access to Wikimedia content and data, in machine-readable formats. This particular wrapper is for the /metrics endpoint.

Wikimedia projects

The package comes bundled with a list of active Wikibooks, Wikinews, Wikipedia, Wikiquote, Wikisource, Wikiversity, Wikivoyage, and Wiktionary wikis.

wikis <- read.csv(system.file("extdata/wikis.csv", package = "waxer"))

The columns are:

language

e.g. "English", "Arabic"

language_code

e.g. "en", "ar"

project_code

e.g. "wikipedia", "wikibooks"

url

e.g. "en.wikipedia.org", "ar.wikibooks.org"

direction

e.g. "left-to-right", "right-to-left"

This dataset was generated from the Wikimedia site matrix.

Metrics

Traffic

wx_top_viewed_articles

Top 1000 viewed articles

wx_unique_device

Unique devices

Editors

wx_active_editors

Number of active editors

wx_top_editors

Top 100 editors per day by number of edits

See also