Generate a rate-limited MediaWiki API caller

Generates a function which makes a request to MediaWiki APIs. This generator is primarily intended for internal use, but can theoretically be used by others to write custom MediaWiki API queries. The generated function is rate-limited at 100 queries per second.

wx_mediawiki_api(project)

Arguments

project
The project for which to generate queryable MW API endpoint

Value

A rate-limited function with parameters:

query

query after endpoint; e.g. action=query&format=json&titles=Main%20page

See also