New pages counts

Number of new pages for a given project and timespan.

wx_new_pages(
  project,
  editor_type = c("all", "anonymous", "user", "bot", "bot group"),
  page_type = c("all", "content", "non-content"),
  granularity = c("daily", "monthly"),
  start_date = "20191101",
  end_date = "20200101"
)

Arguments

project
The name of any Wikimedia project formatted like {language code}.{project name}, for example en.wikipedia. You may pass en.wikipedia.org and the .org will be stripped off. For projects like Wikimedia Commons (without language codes), use commons.wikimedia. For projects like www.mediawiki.org, you can use that full string, or just use mediawiki or mediawiki.org.

editor_type
If you want to filter by editor-type, use one of "anonymous", "group-bot" (registered accounts belonging to the bot group), "name-bot" (registered accounts not belonging to the bot group but having bot-like names) or "user" (registered account not in bot group nor having bot-like name). If you are interested in edits regardless of their editor type, use "all-editor-types" (default).

page_type
If you want to filter by page-type, use one of: "content" (edits made in content namespaces) or "non-content" (edits made in non-content namespaces). If you are interested in editors regardless of their page type, use "all" (default).

granularity
The time unit for the response data. As of today, supported values are daily (default) and monthly.

start_date
The date of the first day to include, in YYYYMMDD format. Can also be a Date or a POSIXt object, which will be auto-formatted.

end_date
The date of the last day to include, in YYYYMMDD format. Can also be a Date or a POSIXt object, which will be auto-formatted.

Granularity

For "monthly" granularity, the start_date and end_date need to contain the full month.

For example, start_date = "20191101" and end_date = "20191231" would only give the monthly total for 2019-11. To include 2019-12, use end_date = "20200101".

Frustratingly, start_date = "20191201" and end_date = "20191231" does not yield 2019-12 monthly total. Use end_date = "20200101" for that.

License

Data retrieved from the API endpoint is available under the CC0 1.0 license.