Top editors by edit count

Top 100 editors per day by number of edits. Caution: anonymous editors IPs are replaced by NA.

wx_top_editors(
  project,
  editor_type = c("all", "anonymous", "user", "bot", "bot group"),
  page_type = c("all", "content", "non-content"),
  start_date = "20191231",
  end_date = "20200101"
)

Arguments

project
The name of any Wikimedia project formatted like {language code}.{project name}, for example en.wikipedia. You may pass en.wikipedia.org and the .org will be stripped off. For projects like Wikimedia Commons (without language codes), use commons.wikimedia. For projects like www.mediawiki.org, you can use that full string, or just use mediawiki or mediawiki.org.

editor_type
If you want to filter by editor-type, use one of "anonymous", "group-bot" (registered accounts belonging to the bot group), "name-bot" (registered accounts not belonging to the bot group but having bot-like names) or "user" (registered account not in bot group nor having bot-like name). If you are interested in edits regardless of their editor type, use "all-editor-types" (default).

page_type
If you want to filter by page-type, use one of: "content" (edits made in content namespaces) or "non-content" (edits made in non-content namespaces). If you are interested in editors regardless of their page type, use "all" (default).

start_date
The date of the first day to include, in YYYYMMDD format. Can also be a Date or a POSIXt object, which will be auto-formatted.

end_date
The date of the last day to include, in YYYYMMDD format. Can also be a Date or a POSIXt object, which will be auto-formatted.

Value

A tibble data frame with the following columns:

project

project

date

Date

user_text

the editor's username; NA if anonymous (instead of IP address)

edits

number of edits the user made on date

rank

the rank based on edits

License

Data retrieved from the API endpoint is available under the CC0 1.0 license.

See also

Examples

if (FALSE) {
wx_top_editors(
  "commons.wikimedia",
  editor_type = "bot",
  page_type = "content",
  start_date = "20191001",
  end_date = "20191031"
)
}