Unique devices

Unique Devices dataset (v1) as developed and maintained by the Wikimedia Foundation since January 1st, 2016. This data set contains the number of unique devices having visited Wikipedia or its sisters project over some period of time. The Analytics team counts unique devices per project per day and month in a way that does not uniquely identify, fingerprint or otherwise track users.

wx_unique_devices(
  project = "mediawiki",
  access_method = c("all", "desktop", "mobile web"),
  granularity = c("daily", "monthly"),
  start_date = "20191231",
  end_date = "20200101"
)

Arguments

project
The name of any Wikimedia project formatted like {language code}.{project name}, for example en.wikipedia. You may pass en.wikipedia.org and the .org will be stripped off. For projects like Wikimedia Commons (without language codes), use commons.wikimedia. For projects like www.mediawiki.org, you can use that full string, or just use mediawiki or mediawiki.org.

access_method
If you want to filter by accessed site, use "desktop" or "mobile". If you are interested in unique devices regardless of accessed site, use "all" (default).

granularity
The time unit for the response data; "daily" by default but if you want to get the top articles of a whole month use "monthly".

start_date
The date of the first day to include, in YYYYMMDD format. Can also be a Date or a POSIXt object, which will be auto-formatted.

end_date
The date of the last day to include, in YYYYMMDD format. Can also be a Date or a POSIXt object, which will be auto-formatted.

Value

A tibble data frame with the following columns:

date

Date; beginning of each month if granularity = "monthly"

devices

Estimated number of unique devices.

offset

The number added to underestimate to produce devices estimate. For more information, refer to the [Nocookie Offset section](https://wikitech.wikimedia.org/wiki/Analytics/Data_Lake/Traffic/Unique_Devices/Last_access_solution#Nocookie_Offset)} \item{underestimate`

The count produced by WMF-Last-Access cookie counting which, by definition, underreports users as we will not be counting users with a fresh session or users browsing without cookies.

Access method

A couple of notes about access_method. First, mobile apps are NOT included in this dataset. access_method = "mobile" refers only to number of unique devices which accessed the mobile website (*.m.* domains).

Second, desktop devices can (and do) access the mobile website, usually because the user clicked on a link shared by a user from a mobile device.

To learn more about how mobile traffic is handled by Wikimedia, refer to mw:Reading/Web/Mobile.

See also