Newly registered users counts

The count of users having registered themselves. This metric excludes user-account created automatically through the auto-login system, and accounts created by peers.

wx_newly_registered_users(
  project,
  granularity = c("daily", "monthly"),
  start_date = "20191101",
  end_date = "20191231"
)

Arguments

project
The name of any Wikimedia project formatted like {language code}.{project name}, for example en.wikipedia. You may pass en.wikipedia.org and the .org will be stripped off. For projects like Wikimedia Commons (without language codes), use commons.wikimedia. For projects like www.mediawiki.org, you can use that full string, or just use mediawiki or mediawiki.org.

granularity
The time unit for the response data. As of today, supported values are daily (default) and monthly.

start_date
The date of the first day to include, in YYYYMMDD format. Can also be a Date or a POSIXt object, which will be auto-formatted.

end_date
The date of the last day to include, in YYYYMMDD format. Can also be a Date or a POSIXt object, which will be auto-formatted.

Value

A tibble data frame with columns

project

project

date

Date; beginning of the month if granularity = "monthly"

new_registered_users

Number of newly registered users

Details

Thanks to CentralAuth, we now are sure that the same usernames in different wiki-projects belong to the same user, and actually, only the first registration is counted in the new registered users metric, since the system then automatically creates user accounts at visit of other projects, and automatic account creations are not counted.

However, for data before CentralAuth, we have no way to know if two accounts with the same name on different wikis belonged to the same user or not. The metric provided for periods before CentralAuth does NOT deduplicate accounts by name across projects, and therefore can be somehow overcounting when used with project-families.

Granularity

For "monthly" granularity, the start_date and end_date need to contain the full month.

For example, start_date = "20191101" and end_date = "20191231" would only give the monthly total for 2019-11. To include 2019-12, use end_date = "20200101".

Frustratingly, start_date = "20191201" and end_date = "20191231" does not yield 2019-12 monthly total. Use end_date = "20200101" for that.

License

Data retrieved from the API endpoint is available under the CC0 1.0 license.

See also

Examples

wx_newly_registered_users(
  "mr.wikipedia", # Marathi
  granularity = "monthly",
  start_date = "20190401",
  end_date = "20191001"
)

#> # A tibble: 6 x 3
#>   project      date       new_registered_users
#>   <chr>        <date>                    <int>
#> 1 mr.wikipedia 2019-04-01                  193
#> 2 mr.wikipedia 2019-05-01                  246
#> 3 mr.wikipedia 2019-06-01                  232
#> 4 mr.wikipedia 2019-07-01                  301
#> 5 mr.wikipedia 2019-08-01                  378
#> 6 mr.wikipedia 2019-09-01                  319