Page view counts

Page view counts

wx_page_views(
  project,
  page_name,
  access_method = c("all", "desktop", "mobile web", "mobile app"),
  agent_type = c("all", "user", "bot"),
  granularity = c("daily", "monthly"),
  start_date = "20191101",
  end_date = "20191231"
)

Arguments

project
The name of any Wikimedia project formatted like {language code}.{project name}, for example en.wikipedia. You may pass en.wikipedia.org and the .org will be stripped off. For projects like Wikimedia Commons (without language codes), use commons.wikimedia. For projects like www.mediawiki.org, you can use that full string, or just use mediawiki or mediawiki.org.

page_name
The title of any article in the specified project. The function takes care of replacing spaces with underscores and URI-encoding, so that non-URI-safe characters like %, / or ? are accepted -- e.g. "Are You the One?" becomes "Are_You_the_One%3F". Internally this is done with a non-exported wx_encode_page_name function. If you need to get the pageviews for multiple pages, you're encouraged to provide all the page names at once as this function has been optimized for that use-case.

access_method
If you want to filter by access method, use one of: "desktop", "mobile app", or "mobile web". If you are interested in pageviews regardless of access method, use "all" (default).

agent_type
If you want to filter by agent type, use "user" or "bot". If you are interested in pageviews regardless of agent type, use "all" (default).

granularity
The time unit for the response data. As of today, supported values are daily (default) and monthly.

start_date
The date of the first day to include, in YYYYMMDD format. Can also be a Date or a POSIXt object, which will be auto-formatted.

end_date
The date of the last day to include, in YYYYMMDD format. Can also be a Date or a POSIXt object, which will be auto-formatted.

Value

A tibble data frame with the following columns:

page_name

the page_name provided by the user

date

Date; beginning of each month if granularity = "monthly"

views

total number of views for the page

License

Data retrieved from the API endpoint is available under the CC0 1.0 license.

See also

Examples

wx_page_views(
  "en.wikipedia",
  c("New Year's Eve", "New Year's Day"),
  start_date = "20191231",
  end_date = "20200101"
)

#> # A tibble: 4 x 3
#>   page_name      date        views
#>   <chr>          <date>      <int>
#> 1 New Year's Day 2019-12-31  61379
#> 2 New Year's Day 2020-01-01 281166
#> 3 New Year's Eve 2019-12-31 202638
#> 4 New Year's Eve 2020-01-01  51715